[feat]: Ladder Interface 생성
